refactor(runtime): split runtime into memory, string, and IO modules
